Verify that callbacks don't leak strings (#2435)
* Clean up callbacks that call from native into managed with strings

* Replace some byte* arguments with RealmValue + remove a bunch of unnecessary unsafe-s

* Update to latest v11

+++ wrappers/src/app_cs.cpp
@@ -44,7 +44,7 @@ using LogMessageCallbackT = void(void* managed_handler, realm_value_t message, u
 using UserCallbackT = void(void* tcs_ptr, SharedSyncUser* user, MarshaledAppError err);
 using VoidCallbackT = void(void* tcs_ptr, MarshaledAppError err);
 using BsonCallbackT = void(void* tcs_ptr, realm_value_t response, MarshaledAppError err);
-using ApiKeysCallbackT = void(void* tcs_ptr, UserApiKey* api_keys, int api_keys_len, MarshaledAppError err);
+using ApiKeysCallbackT = void(void* tcs_ptr, UserApiKey* api_keys, size_t api_keys_len, MarshaledAppError err);
 
 namespace realm {
     namespace binding {
